Properly propagate Selection Lists, fixes aliases
We should propagate the selection list down to the children elements,
otherwise aliases would be lost and the query wouldn't be properly
implemented.

While the fix is in `sgqlc/types/__init__.py`, the test is in
`sgqlc/operation/__init__.py` as it deals with results interpretation
and selection lists.

Fixes #16