fix auto selection of fields, adds depth limit
This commit fixes the auto-selection of fields when no fields of a
container type is explicitly selected:

 - keep the trail of types, do not include fields leading to loops;
 - limit the recursion to a given depth, defaults to 2

The auto-selection is mostly a nice to have feature to allow simple
types to be selected without the need of an empty `__fields__()` and
the likes. However in complex scenarios such as the GitHub API, this
leads to huge queries for most types, then the limit to 2 nesting
levels is reasonable. The user may change that with
`auto_select_depth`.

Closes #17