Add inline fragment support, enhances Interface and Union usage
Introduce the `__as__(ActualType)` method to `Selection` and
`SelectionList`, allows to cast to `ActualType` if the returned
`__typename` matches that, extending the field selection.

The `__typename` is automatically selected if `__as__()` is used, as
well as it's used to interpret the returned JSON data with the
specific type.

Currently the code does no verification if the field type is
implemented (interfaces) or contains the option (union) of the
type given to `__as__()`.

The operation outputs an Inline Fragment `... on ActualType {}`.

Closes #3
Closes #18