codegen: fix handling of argument's default value
The default value doesn't come as JSON serialized object as I first
understood, it's a GraphQL DSL with the definition.

As parsing the DSL ourselves for a simple case should be simple,
trying to do it may lead to new issuess in the future, then we're not
dependent on the `graphql-core` package in order to use their parser
and AST Visitor helper.

Fixes #22