endpoint: more robust error handling
The data format may be not in the GraphQL error object, then adjust
the object to match that.

For instance the data doesn't need to be a dict, it could be just a
string, in this case it's not interesting to us (report only an HTTP
error). See bug #24

For JSON objects that do have an `errors` key, convert that to a
GraphQL error object shape (if needed) and also remember the raised
exception.

Last but not least, do not trust the `content-type` for json,
try-except around the decoder and fail gracefully.

Fixes #24