pre-commit only do lint on changed files, only test if python files changed
This speeds up the commit checks a bit only linting the changed files
(or all files if configuration/setup.* files changed).

The nosetests are only executed if python files (or config) changed,
but the test is executed for the whole set as doctests change
depending on the other files being imported, they affect the global
schema.