[Fix][profusion/sgqlc#35] Don't error when populating fields from casted list
When casting using `__as__`, we store the cast-to type in a `__casts__` list.
When constructing the response to a query, we check whether `__casts__` is
populated, and if so we attempt to retrieve the `__typename` field from the
JSON object in question. If that object is a list, it doesn't have fields so the
`get()` call fails.

For an example piece of failing code, see profusion/sgqlc#35.

This fix catches the `AttributeError` raised when the `get()` call fails, and
ignores it. The elements within the list will be separately cast to the correct
type recursively.

It also adds tests which call `__as__` on a list and then construct the response
to ensure it doesn't fail.