Support canceling queries and reusing websockets
Move the generator implementation into a dedicated class
to allow for canceling of running queries.  Also support
passing an already connected websocket to the endpoint
to allow multiple queries to happen on the same connection.

Note: in order to support multiple concurrent connections,
this probably needs to move to using websocket-client's
asynchronous WebSocketApp instead of synchronously
calling websocket.create_connection().recv(), and ensuring
the notification messages get routed to the correct
generator.