Change version requirement to allow Python 3.5
There are a number of test failures remaining, so this is perhaps
somewhat optimistic. However, all remaining failures appear to arise
from differences in ordering of fields or arguments when printing
objects as GraphQL (presumably related to the lack of `dict` ordering
guarantees in Python 3.5), which should not break very much in practice.