codegen: add operations generator
This is another helper that given a GraphQL domain-specific-language
(DSL), it will generate the Python code using SGQL types.

Since fragments are not supported by sgqlc.operation, it will inline
the fields. Eventually the named fragment support will be added.

Variables are respected, that allows parameters to be given to
operations.

A simplified validation is done, it's not fully complete (one can use
existing tools to validate `.gql` files against a schema), but should
serve most people, including:
 - input type mismatch
 - invalid selections/fields
 - invalid types
 - invalid arguments
 - missing required arguments
 - invalid enum values

In the future I plan to add support to generate typings, so it will
help mypy users.