codegen-operation: fix argument name validation
This should not cause a regression against #123, I've tested a similar
query and go correct results.

This basically reverts #123. The other attempt to fix it, using `var`
if it was given makes no sense (my bad!) since variable name can be
anything, it doesn't need to match the argument name.

Closes: #133
Closes: #144