codegen-operation: improved validation and variable handling
Rewrite most of the variable handling, now the variable definitions
validate the type and the default value, including if there is a
non-null type with a default value (which is not accepted by most
GraphQL servers) and if they were not used (also rejected by some
servers).

Using variables will also check if their types matches with the actual
usage, otherwise it will fail.

Last but not least, the validation errors are now properly captured
and converted into `SystemExit` pointing to the node location.

This also fixes validation of null values with nullable fields.

The format_typename_usage() was changed to use the shorter `_schema`
symbol.