Fix construction of list types with already-realized types
As noticed by @wolfgangpfnuer in #173, the _create_list_of_wrapper()
implementation of realize_type() does not work if the given value is
already a realized type or a variable.

That is done in _create_non_null_wrapper() version, to avoid
inconsistencies, isolate the realize_type in a factory and reuse it in
both places.

Add a test to check the issue is solved and avoid future breakages.

Closes: #173
Closes: #166