Python 3.13 support + update locked dependencies + coverallsapp/github-action
Coveralls is stuck with <3.13, but it's only used to report the collected
values, then use their new coverallsapp/github-action instead, it will
run on its own python container, whatever version/deps it needs.

This allows us to run on 3.13, which will be the primary development
version from now on. Then update the poetry-locked dependencies to
their newest versions.

GitHub action: do not check 3.8 since it's EOL and newest black is
incompatible.

Fixes: #247